+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [2.2.2]
+
+- Introduce `state-flow.assertions.matcher-combinators/match?`
+  - deprecate `state-flow.cljtest/match?`
+  - add `state-flow.refactoring-tools.refactor-match/refactor-all` to help with
+    refactoring to the new version
+
 ## [2.2.1]
 - Use vectors in internal state data structure instead of cats pairs